+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `error` function
 - `hash` function
 - `try` function
+- `QColor.mag` function
 
 ### Changed
 - Fixed function subcontexts setting the wrong variables as arguments due to a parent context having another variable named the same way.